chore(infra): fix rm code in cleaning autogen proto code

+++ scripts/clean_proto.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-rm $(find internal/gen/ -type f -name '*.pb.go')
+if [[ ! -z $(find internal/gen/ -type f -name '*.pb.go') ]]; then
+	rm $(find internal/gen/ -type f -name '*.pb.go')
+fi
